Colocate type definitions into one file
