--- app/types/Account.ts
@@ -1,7 +1,0 @@
-export type AccountId = number;
-
-export type Account = {
-  account_id: AccountId;
-  limit: number;
-  products: string[];
-};
