--- app/types/Customer.ts
@@ -1,10 +1,0 @@
-import { AccountId } from "./Account";
-
-export type Customer = {
-  username: string;
-  name: string;
-  address: string;
-  birthdate: string;
-  email: string;
-  accounts: AccountId[];
-};
